requirements.txt
google-genai
langchain~=0.3.20
python-dotenv~=1.0.1
langchain-google-genai~=2.1.0
